requirements.txt
matplotlib==3.7.2
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
scikit_learn==1.3.0
scipy==1.11.4
seaborn==0.13.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
sphinx
sphinx-rtd-theme
